base/view, urls account/view, signals, url

+++ account/views.py
@@ -1,42 +1,32 @@
-from django.shortcuts import render
 from rest_framework import status
-from rest_framework.decorators import api_view
-from django.contrib.auth import authenticate, login, logout
 from rest_framework.response import Response
 from rest_framework.views import APIView
-import base.views
-from django.http import HttpResponse, HttpResponseRedirect
-from django.urls import reverse
-from account.models import *
-from django.contrib.auth.decorators import login_required
 from .signals import *
-
 from .serializers import *
 from django.db import IntegrityError
-from app.authenticate import Authenticate
 
 
 class AutheView(APIView):
-    # permission_classes = (Authenticate,)
-
     def get(self, request, user_id=None):
         if request.user.is_staff:
-            if user_id is None:
+            user = User.objects.filter(id=user_id)
+            if user.count() == 0:
                 return Response({"message": "User with this profile was not found!"},
                                 status=status.HTTP_400_BAD_REQUEST)
-
             user = Profile.objects.get(user_id=user_id)
             user = ProfileSerializer(user).data
-
             return Response(user, status=status.HTTP_302_FOUND)
         else:
-            user = Profile.objects.get(user_id=request.user.id)
-            user = ProfileSerializer(user).data
-            context = {
-                "massege": "access denied",
-                "your profile": user
-            }
-            return Response(context, status=status.HTTP_400_BAD_REQUEST)
+            try:
+                user = Profile.objects.get(user_id=request.user.id)
+                user = ProfileSerializer(user).data
+                context = {
+                    "massege": "access denied",
+                    "your profile": user
+                }
+                return Response(context, status=status.HTTP_400_BAD_REQUEST)
+            except:
+                return Response({"massege": "permission error"}, status=status.HTTP_203_NON_AUTHORITATIVE_INFORMATION)
 
     def post(self, request):
         try:
@@ -50,21 +40,28 @@ class AutheView(APIView):
             return Response({'error': str(err)}, status=status.HTTP_400_BAD_REQUEST)
 
     def put(self, req):
+        # permission_classes = (Authenticate,)
         user = User.objects.filter(id=req.user.id).update(**req.data)
         user = User.objects.get(id=req.user.id)
         user = UserSerializer(user).data
 
         return Response(user, status=status.HTTP_200_OK)
 
-    def delete(self, request):
+    def delete(self, request, user_id):
         if request.user.is_staff:
-            try:
-                user = User.objects.get(id=request.data["id"])
+            userid = User.objects.filter(id=user_id)
+            if userid.count() > 0:
+                user = User.objects.get(id=user_id)
                 user.delete()
-                return Response({"message": "user {} deleted successfully".format(request.data["id"])},
+                return Response({"message": "user {} deleted successfully".format(user_id)},
                                 status=status.HTTP_200_OK)
-            except:
-                return Response({"message": "user {} not exist".format(request.data["id"])},
-                                status=status.HTTP_400_BAD_REQUEST)
+            else:
+                user = User.objects.all()
+                user = UserSerializer(user, many=True).data
+                data = {
+                    "massege": "user id:{}, was not found!".format(user_id),
+                    "existdata": user
+                }
+                return Response(data, status=status.HTTP_409_CONFLICT)
         else:
             return Response({"message": "access denied"}, status=status.HTTP_400_BAD_REQUEST)
